etl/requirements.txt
# Portland Commons Dashboard — ETL Pipeline Dependencies
# Python 3.12+

# HTTP & API
requests>=2.31,<3
httpx>=0.27,<1

# Database
psycopg2-binary>=2.9,<3
sqlalchemy>=2.0,<3

# Data processing
pandas>=2.2,<3
numpy>=1.26,<3

# HTML/XML parsing
beautifulsoup4>=4.12,<5
lxml>=5.1,<6

# Scheduling
schedule>=1.2,<2

# Environment & config
python-dotenv>=1.0,<2

# Census API
census>=0.8,<1
us>=3.1,<4

# GTFS parsing
gtfs-kit>=8.0,<9

# Geospatial (optional, for PostGIS work)
geopandas>=0.14,<1
shapely>=2.0,<3

# Data validation
pandera>=0.18,<1

# Logging & monitoring
structlog>=24.1,<25

# CLI
click>=8.1,<9

# Testing
pytest>=8.0,<9
pytest-asyncio>=0.23,<1
